test: move the Plex tests out of the release-info file
Cloud review nit on #202 (non-blocking, but correct): TestPlexErrorFormatting
covers the plex.tv token-fetch fixes, not createFromSearch() failing to
populate release metadata, so it matched neither the file's name nor its
docstring.

Split into tests/unit/test_plex_server.py with a docstring covering both
crashes in that path -- the dead ex() helper and the unbound `response` --
and why they were nasty (both sit inside error handling, so they replaced the
real diagnostic at exactly the wrong moment).
